+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 0.8.5
+* Enhancements
+  * Add colored log output
+  * Add telemetry support for execution
+
 ## 0.8.4 (2022-03-09)
 * Bugfix
   * add missing excludes from dump function (collections)